Struct ra_ide_api::Analysis

pub struct Analysis { /* fields omitted */ }

Analysis is a snapshot of a world state at a moment in time. It is the main entry point for asking semantic information about the world. When the world state is advanced using AnalysisHost::apply_change method, all existing Analysis are canceled (most method return Err(Canceled)).

Methods

impl Analysis

pub fn from_single_file(text: String) -> (Analysis, FileId)

pub fn status(&self) -> String

Debug info about the current state of the analysis

pub fn file_text(&self, file_id: FileId) -> Arc<String>

Gets the text of the source file.

pub fn parse(&self, file_id: FileId) -> TreeArc<SourceFile>

Gets the syntax tree of the file.

pub fn file_line_index(&self, file_id: FileId) -> Arc<LineIndex>

Gets the file's LineIndex: data structure to convert between absolute offsets and line/column representation.

pub fn extend_selection(&self, frange: FileRange) -> Cancelable<TextRange>

Selects the next syntactic nodes encompassing the range.

pub fn matching_brace(&self, position: FilePosition) -> Option<TextUnit>

Returns position of the matching brace (all types of braces are supported).

pub fn syntax_tree(
&self,
file_id: FileId,
text_range: Option<TextRange>
) -> String

Returns a syntax tree represented as String, for debug purposes.

pub fn join_lines(&self, frange: FileRange) -> SourceChange

Returns an edit to remove all newlines in the range, cleaning up minor stuff like trailing commas.

pub fn on_enter(&self, position: FilePosition) -> Option<SourceChange>

Returns an edit which should be applied when opening a new line, fixing up minor stuff like continuing the comment.

pub fn on_eq_typed(&self, position: FilePosition) -> Option<SourceChange>

Returns an edit which should be applied after = was typed. Primarily, this works when adding let =.

pub fn on_dot_typed(&self, position: FilePosition) -> Option<SourceChange>

Returns an edit which should be applied when a dot ('.') is typed on a blank line, indenting the line appropriately.

pub fn file_structure(&self, file_id: FileId) -> Vec<StructureNode>

Returns a tree representation of symbols in the file. Useful to draw a file outline.

pub fn folding_ranges(&self, file_id: FileId) -> Vec<Fold>

Returns the set of folding ranges.

pub fn symbol_search(&self, query: Query) -> Cancelable<Vec<NavigationTarget>>

Fuzzy searches for a symbol.

pub fn goto_definition(
&self,
position: FilePosition
) -> Cancelable<Option<RangeInfo<Vec<NavigationTarget>>>>

pub fn goto_implementation(
&self,
position: FilePosition
) -> Cancelable<Option<RangeInfo<Vec<NavigationTarget>>>>

pub fn goto_type_definition(
&self,
position: FilePosition
) -> Cancelable<Option<RangeInfo<Vec<NavigationTarget>>>>

pub fn find_all_refs(
&self,
position: FilePosition
) -> Cancelable<Option<ReferenceSearchResult>>

Finds all usages of the reference at point.

pub fn hover(
&self,
position: FilePosition
) -> Cancelable<Option<RangeInfo<HoverResult>>>

Returns a short text describing element at position.

pub fn call_info(&self, position: FilePosition) -> Cancelable<Option<CallInfo>>

Computes parameter information for the given call expression.

pub fn parent_module(
&self,
position: FilePosition
) -> Cancelable<Vec<NavigationTarget>>

Returns a mod name; declaration which created the current module.

pub fn crate_for(&self, file_id: FileId) -> Cancelable<Vec<CrateId>>

Returns crates this file belongs too.

pub fn crate_root(&self, crate_id: CrateId) -> Cancelable<FileId>

Returns the root file of the given crate.

pub fn runnables(&self, file_id: FileId) -> Cancelable<Vec<Runnable>>

Returns the set of possible targets to run for the current file.

pub fn highlight(&self, file_id: FileId) -> Cancelable<Vec<HighlightedRange>>

Computes syntax highlighting for the given file.

pub fn highlight_as_html(
&self,
file_id: FileId,
rainbow: bool
) -> Cancelable<String>

Computes syntax highlighting for the given file.

pub fn completions(
&self,
position: FilePosition
) -> Cancelable<Option<Vec<CompletionItem>>>

Computes completions at the given position.

pub fn assists(&self, frange: FileRange) -> Cancelable<Vec<Assist>>

Computes assists (aka code actions aka intentions) for the given position.

pub fn diagnostics(&self, file_id: FileId) -> Cancelable<Vec<Diagnostic>>

Computes the set of diagnostics for the given file.

pub fn type_of(&self, frange: FileRange) -> Cancelable<Option<String>>

Computes the type of the expression at the given position.

pub fn rename(
&self,
position: FilePosition,
new_name: &str
) -> Cancelable<Option<SourceChange>>

Returns the edit required to rename reference at the position to the new name.